requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu  # or use faiss-gpu if you're using GPU
pypdf
